zsh: alias ssh to use TERM=xterm-256color env
for some reason alacritty (my terminal emulator) is not recognized when
I try to ssh in some other machine, causing somethings to not work (like
the `clear` command for example).

This env variable tricks the remote system to think I am using other
terminal emulator. Probably not the best solution, but it is working
just fine until now.